+++ src/components/drawer/__tests__/index.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { shallow, ShallowWrapper } from "enzyme";
 
 import Button from "components/button";
-import { availableWidgets } from "widgets";
+import widgets from "widgets";
 
 import Drawer from "../index";
 
@@ -14,6 +14,6 @@ describe("<Drawer />", () => {
   });
 
   it("renders a button per available ", () => {
-    expect(wrapper.find(Button)).toHaveLength(availableWidgets.length);
+    expect(wrapper.find(Button)).toHaveLength(Object.keys(widgets).length);
   });
 });