+++ src/widgets/list.ts
@@ -1,0 +1,6 @@
+import { Widgets } from "./index";
+export default {
+  "date-time": { configurable: false },
+  search: { configurable: true },
+  text: { configurable: false }
+} as Widgets;