+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- :gem: List available widgets dynamically.
 
 ## 1.0.9 (2019-09-30)